requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
scrapy==2.11.0
google-cloud-storage==2.13.0
google-auth==2.23.4
crochet==2.1.1
python-dotenv==1.0.0
pydantic==2.5.1